+++ src/caselawclient/models/search_results.py
@@ -1,3 +1,4 @@
+import os
 from typing import List
 
 from lxml import etree
@@ -50,3 +51,29 @@ class SearchResults:
             "//search:response/search:result", namespaces=self.NAMESPACES
         )
         return [element for element in results]
+
+
+class SearchMatches:
+    """
+    Represents a collection of search matches obtained from XML data.
+    """
+
+    def __init__(self, xml: str):
+        """
+        Initialize a SearchMatches instance with XML data.
+
+        Args:
+            xml (str): XML data as a string.
+        """
+        self._root = etree.fromstring(xml)
+
+    def transform_to_html(self) -> str:
+        """
+        Transform the XML data to HTML using XSLT.
+
+        Returns:
+            str: The transformed HTML content.
+        """
+        file_path = os.path.join(os.path.dirname(__file__), "search_match.xsl")
+        xslt_transform = etree.XSLT(etree.parse(file_path))
+        return str(xslt_transform(self._root))